third_party/blink/web_tests/external/wpt/mathml/presentation-markup/scripts/underover-parameters-4.tentative-expected.txt

This is a testharness.js-based test.
[FAIL] AccentBaseHeight, UnderbarVerticalGap
  assert_approx_equals: munderover: accent over short base expected 40 +/- 2 but got 30
[FAIL] AccentBaseHeight, OverbarExtraAscender
  assert_approx_equals: mover: accent over short base expected 40 +/- 2 but got 30
[FAIL] AccentBaseHeight, OverbarVerticalGap
  assert_approx_equals: mover: accent over short base expected 40 +/- 2 but got 140
Harness: the test ran to completion.
